AP_NavEKF3: bugfix commit  and some docs update and wip harmless other changes
suggestion from peter

bugfixes

buzz plausibe hacking

classic esp periph comments only

AP_Compass: honour AP_COMPASS_ENABLED more

esp32: disable a bunch of subsystems on Periph or it wont link

AP_Periph:  make can driver  incapable of dividing by zero

compile error on esp32 periph discovered this potential issue with HAL_NUM_CAN_IFACES=0

AP_Scripting: lua binding linker error when building on esp32 periph

important esp32 py fixes

esp32 periph, support linking non-s3 and s3 periphs

+++ libraries/AP_NavEKF3/AP_NavEKF3.cpp
@@ -810,8 +810,8 @@ bool NavEKF3::InitialiseFilter(void)
         }
 
         // check if there is enough memory to create the EKF cores
-        uint16_t x = AP::dal().available_memory();
-        printf("ekf3 stats: avail:%d numcores:%d sizeof:%d\n",x,num_cores,sizeof(NavEKF3_core)*num_cores + 4096);
+        uint32_t x = AP::dal().available_memory();
+        printf("ekf3 stats: avail:%lu numcores:%d sizeof:%lu\n",(unsigned long)x,num_cores,(unsigned long)(sizeof(NavEKF3_core)*num_cores + 4096));
         if (x < (sizeof(NavEKF3_core)*num_cores + 4096)) {
             GCS_SEND_TEXT(MAV_SEVERITY_CRITICAL, "EKF3 not enough memory: %d",(unsigned int)x);
             _enable.set(0);
